Partially fixed invalid float points

+++ kernel/apps/calc.h
@@ -16,15 +16,15 @@
 // math storage variables
 char calcInput[CALCSIZE + 2]; // For `+0`
 int mathOp[CALCSIZE];
-double tempNum;
-double strNum[CALCSIZE];
+float tempNum;
+float strNum[CALCSIZE];
 int strNumCount;
 
 // Initializes the variable map! IMPORTANT!
 void initialize_calc();
 
 // contatination function for the calculator app
-double concat(double, double);
+float concat(float, float);
 
 // main calc() function to run from kernel.c
 void calc(string);
